+++ src/executor/executor.cc
@@ -1,9 +1,44 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-#include "executor.hpp"
-
+#include <mscclpp/executor.hpp>
+#include <mscclpp/proxy_channel.hpp>
+#include <mscclpp/sm_channel.hpp>
 #include <set>
+
+#include "execution_plan.hpp"
+
+namespace mscclpp {
+struct ExecutionContextKey {
+  void* sendBuff;
+  void* recvBuff;
+  size_t sendBuffSize;
+  size_t recvBuffSize;
+  std::string plan;
+
+  bool operator==(const ExecutionContextKey& other) const {
+    return sendBuff == other.sendBuff && recvBuff == other.recvBuff && sendBuffSize == other.sendBuffSize &&
+           recvBuffSize == other.recvBuffSize && plan == other.plan;
+  }
+};
+}  // namespace mscclpp
+
+namespace std {
+template <>
+struct hash<std::pair<mscclpp::BufferType, int>> {
+  std::size_t operator()(const std::pair<mscclpp::BufferType, int>& key) const {
+    return std::hash<int>()(key.second) ^ std::hash<int>()(static_cast<int>(key.first));
+  }
+};
+
+template <>
+struct hash<mscclpp::ExecutionContextKey> {
+  std::size_t operator()(const mscclpp::ExecutionContextKey& key) const {
+    return std::hash<void*>()(key.sendBuff) ^ std::hash<void*>()(key.recvBuff) ^ std::hash<size_t>()(key.sendBuffSize) ^
+           std::hash<size_t>()(key.recvBuffSize) ^ std::hash<std::string>()(key.plan);
+  }
+};
+}  // namespace std
 
 namespace {
 static const mscclpp::Transport IBs[] = {mscclpp::Transport::IB0, mscclpp::Transport::IB1, mscclpp::Transport::IB2,
@@ -13,8 +48,37 @@ static const mscclpp::Transport IBs[] = {mscclpp::Transport::IB0, mscclpp::Trans
 
 namespace mscclpp {
 
-Executor::Executor(std::shared_ptr<Communicator> comm, const std::unordered_map<int, mscclpp::Connection> connections)
-    : impl_(std::make_shared<Impl>(comm, connections)) {}
+struct ExecutionContext {
+  std::unordered_map<std::pair<BufferType, int>, mscclpp::RegisteredMemory> registeredMemories;
+  std::vector<std::shared_ptr<mscclpp::SmDevice2DeviceSemaphore>> smSemaphores;
+  std::vector<mscclpp::SemaphoreId> proxySemaphores;
+  std::vector<mscclpp::SmChannel> smChannels;
+  std::vector<mscclpp::SimpleProxyChannel> proxyChannels;
+  std::vector<DeviceExecutionPlan> deviceExecutionPlans;
+  std::shared_ptr<char> scratchBuffer;
+  size_t scratchBufferSize;
+};
+
+struct Executor::Impl {
+  std::shared_ptr<Communicator> comm;
+  const std::unordered_map<int, std::shared_ptr<Connection>> connections;
+  std::shared_ptr<ProxyService> proxyService;
+  std::unordered_map<ExecutionContextKey, ExecutionContext> contexts;
+
+  Impl(std::shared_ptr<Communicator> comm, const std::unordered_map<int, std::shared_ptr<Connection>> connections);
+  ExecutionContext setupExecutionContext(int rank, void* sendbuff, void* recvbuff, size_t sendBufferSize,
+                                         size_t recvBufferSize, const ExecutionPlan& plan);
+  void setupRegisteredMemories(ExecutionContext& context, void* sendbuff, void* recvbuff, size_t sendBufferSize,
+                               size_t recvBufferSize, int rank, const ExecutionPlan& plan);
+  void setupChannels(ExecutionContext& context, void* sendbuff, void* recvbuff, size_t sendBufferSize, int rank,
+                     const ExecutionPlan& plan);
+  void launchKernel(ExecutionContext& context);
+  ~Impl() = default;
+};
+
+Executor::Executor(std::shared_ptr<Communicator> comm,
+                   const std::unordered_map<int, std::shared_ptr<Connection>> connections)
+    : impl_(std::make_unique<Impl>(comm, connections)) {}
 
 void Executor::execute(void* sendbuff, void* recvBuff, size_t sendBuffSize, size_t recvBuffSize,
                        const ExecutionPlan& plan) {
@@ -31,12 +95,12 @@ Executor::Impl::Impl(std::shared_ptr<Communicator> comm,
 
 ExecutionContext Executor::Impl::setupExecutionContext(int rank, void* sendbuff, void* recvbuff, size_t sendBufferSize,
                                                        size_t recvBufferSize, const ExecutionPlan& plan) {
-  ExecutionContextKey key = {sendbuff, recvbuff, sendBufferSize, recvBufferSize, plan.getName()};
+  ExecutionContextKey key = {sendbuff, recvbuff, sendBufferSize, recvBufferSize, plan.impl_->name};
   if (this->contexts.find(key) != this->contexts.end()) {
     return this->contexts[key];
   }
   ExecutionContext context;
-  size_t scratchBufferSize = plan.getScratchBufferSize(rank, sendBufferSize);
+  size_t scratchBufferSize = plan.impl_->getScratchBufferSize(rank, sendBufferSize);
   std::shared_ptr<char> scratchBuffer = allocExtSharedCuda<char>(scratchBufferSize);
   context.scratchBuffer = scratchBuffer;
   context.scratchBufferSize = scratchBufferSize;
@@ -48,7 +112,7 @@ ExecutionContext Executor::Impl::setupExecutionContext(int rank, void* sendbuff,
 void Executor::Impl::setupRegisteredMemories(ExecutionContext& context, void* sendbuff, void* recvbuff,
                                              size_t sendBufferSize, size_t recvBufferSize, int rank,
                                              const ExecutionPlan& plan) {
-  int nranksPerNode = plan.nranksPerNode();
+  int nranksPerNode = plan.impl_->nranksPerNode;
   auto getTransportFlags = [&](std::vector<ChannelInfo>& infos, int rank) {
     TransportFlags flags;
     for (ChannelInfo& info : infos) {
@@ -82,9 +146,9 @@ void Executor::Impl::setupRegisteredMemories(ExecutionContext& context, void* se
     return std::vector<int>(peers.begin(), peers.end());
   };
 
-  std::vector<BufferType> bufferTypes = plan.getConnectedBufferTypes(rank);
+  std::vector<BufferType> bufferTypes = plan.impl_->getConnectedBufferTypes(rank);
   for (BufferType bufferType : bufferTypes) {
-    std::vector<ChannelInfo> channelInfos = plan.getChannelInfos(rank, bufferType);
+    std::vector<ChannelInfo> channelInfos = plan.impl_->getChannelInfos(rank, bufferType);
     TransportFlags transportFlags = getTransportFlags(channelInfos, rank);
     RegisteredMemory memory =
         this->comm->registerMemory(getBufferInfo(bufferType).first, getBufferInfo(bufferType).second, transportFlags);
@@ -95,7 +159,7 @@ void Executor::Impl::setupRegisteredMemories(ExecutionContext& context, void* se
       remoteRegMemoryFutures.push_back(comm->recvMemoryOnSetup(peer, 0));
     }
     comm->setup();
-    for (int i = 0; i < remoteRegMemoryFutures.size(); i++) {
+    for (size_t i = 0; i < remoteRegMemoryFutures.size(); i++) {
       context.registeredMemories[{bufferType, connectedPeers[i]}] = std::move(remoteRegMemoryFutures[i].get());
     }
   }
@@ -107,7 +171,7 @@ void Executor::Impl::setupChannels(ExecutionContext& context, void* sendbuff, vo
   std::vector<std::shared_ptr<SmDevice2DeviceSemaphore>> smSemaphores;
   std::vector<mscclpp::SemaphoreId> proxySemaphores;
   for (ChannelType channelType : channelTypes) {
-    std::vector<ChannelInfo> channelInfos = plan.getChannelInfos(rank, channelType);
+    std::vector<ChannelInfo> channelInfos = plan.impl_->getChannelInfos(rank, channelType);
     for (ChannelInfo& info : channelInfos) {
       for (int peer : info.connectedPeers) {
         if (channelType == ChannelType::SM) {
@@ -135,11 +199,10 @@ void Executor::Impl::setupChannels(ExecutionContext& context, void* sendbuff, vo
     }
   };
   for (ChannelType channelType : channelTypes) {
-    std::vector<ChannelInfo> channelInfos = plan.getChannelInfos(rank, channelType);
+    std::vector<ChannelInfo> channelInfos = plan.impl_->getChannelInfos(rank, channelType);
     int index = 0;
     for (ChannelInfo& info : channelInfos) {
       void* src = getBuffer(info.srcBufferType);
-      void* dst = getBuffer(info.dstBufferType);
       TransportFlags transport = context.registeredMemories.begin()->second.transports();
       RegisteredMemory localMemory = this->comm->registerMemory(src, sendBufferSize, transport);
       for (int peer : info.connectedPeers) {
@@ -159,7 +222,8 @@ void Executor::Impl::setupChannels(ExecutionContext& context, void* sendbuff, vo
 
 void Executor::Impl::launchKernel(ExecutionContext& context) {
   // Need to change to use flush function and make sure the proxy service will get the latest data.
-  this->proxyService->startProxy();
+  // may need atomic variable
+  // this->proxyService->startProxy();
 }
 
 }  // namespace mscclpp
